travis: temporarily disable subtree check on libsecp
I did not format my commit messages correctly when updating libsecp-zkp, so
there is no way that this check will pass on the 0.21 rebase PR, unless I
rewrite those commits to have include lists of all secp-zkp commits that
were pulled in.

For now it should just be manually verified that I updated the secp subtree
correctly. In a later PR we should re-enable this check and in future we
should copy the Core workflow (which to the best of my knowledge, we never
have done) when updating the -zkp subtree.