wallet: don't reserve change destinations for freshly issued assets
In our coin selection logic we're reserving change destinations for
newly issued assets, but we always issue exactly as much as is
needed, so there is guaranteed to be no change outputs for these
assets.

By reserving them we run the risk of returning "keypool empty" errors
when we don't actually need anything from the keypool.