wallet: treat change for preset-input assets the same as that for destination assets
In Bitcoin Core #22008 we refactored the way that we generate change
addresses to allow address generation to fail provided that the change
is eventually dropped. We updated our code to follow in a5d97b363b51
but only updated the logic for explicit destinations. In fact, we had
a second set of change-address logic to create change addresses for
preset inputs, and we did not update this.

Unify these things to reduce the bug surface.