checkpoint: have fuzzer working using static library
This is a rollup of a dozen or so commits which iterated on the fuzzer.
This incarnation uses a static library to link into Elements, which is
compiled using some crazy hacks using `cargo fuzz` to get all the
libfuzzer instrumentation in place.

It gets high coverage but cannot generate pegin data or pegouts or
annexes (the latter two because of a bug producing Vecs where the
initial byte is always less than 16.)